Added the formatter.

+++ api/src/main/java/org/victor5171/revoluttest/api/LatestRatesResponse.kt
@@ -5,5 +5,5 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class LatestRatesResponse(
     val baseCurrency: String,
-    val rates: Map<String, Float>
+    val rates: Map<String, Double>
 )
