+++ feature-rateconversion/src/main/java/org/victor5171/revoluttest/rateconversion/viewmodel/ConvertedRate.kt
@@ -3,5 +3,5 @@ package org.victor5171.revoluttest.rateconversion.viewmodel
 data class ConvertedRate(
     val currencyIdentifier: String,
     val currencyName: String,
-    val value: Float
+    val value: Double
 )
