+++ persistence/src/main/java/org/victor5171/revoluttest/persistence/rate/RateDTO.kt
@@ -2,5 +2,5 @@ package org.victor5171.revoluttest.persistence.rate
 
 data class RateDTO(
     val destinationCurrency: String,
-    val multiplier: Float
+    val multiplier: Double
 )
