+++ repository/src/main/java/org/victor5171/revoluttest/repository/rateconversion/Rate.kt
@@ -2,5 +2,5 @@ package org.victor5171.revoluttest.repository.rateconversion
 
 data class Rate(
     val destinationCurrency: String,
-    val multiplier: Float
+    val multiplier: Double
 )
