+++ room-persistence/src/main/java/org/victor5171/revoluttest/persistence/room/rate/RateMinimal.kt
@@ -7,5 +7,5 @@ internal data class RateMinimal(
     val destinationCurrency: String,
 
     @ColumnInfo(name = "multiplier")
-    val multiplier: Float
+    val multiplier: Double
 )
